requirements.txt
docopt==0.6.2
facebook-sdk==2.0.0
pysolr==3.5.0
python-dateutil==2.4.2
requests==2.11.1
six==1.10.0
urllib3==1.14
voluptuous==0.9.3
